add kafka instrumentation
