+++ lib/vector-core/core-common/src/internal_event/mod.rs
@@ -1,4 +1,8 @@
-use metrics::counter;
+mod bytes_sent;
+mod events_sent;
+
+pub use bytes_sent::BytesSent;
+pub use events_sent::EventsSent;
 
 pub trait InternalEvent {
     fn emit_logs(&self) {}
@@ -43,30 +47,4 @@ pub fn emit(event: &impl InternalEvent) {
 pub fn emit(event: &impl InternalEvent) {
     event.emit_logs();
     event.emit_metrics();
-}
-
-// Common Events
-
-#[derive(Debug)]
-pub struct EventsSent {
-    pub count: usize,
-    pub byte_size: usize,
-}
-
-impl InternalEvent for EventsSent {
-    fn emit_logs(&self) {
-        trace!(message = "Events sent.", count = %self.count, byte_size = %self.byte_size);
-    }
-
-    fn emit_metrics(&self) {
-        if self.count > 0 {
-            // events_out_total is emitted by `Acker`
-            counter!("component_sent_events_total", self.count as u64);
-            counter!("component_sent_event_bytes_total", self.byte_size as u64);
-        }
-    }
-
-    fn name(&self) -> Option<&str> {
-        Some("EventsSent")
-    }
 }
