+++ website/cue/reference/components/sinks/kafka.cue
@@ -164,6 +164,9 @@ components: sinks: kafka: {
 	how_it_works: components._kafka.how_it_works
 
 	telemetry: metrics: {
+		component_sent_events_total:         components.sources.internal_metrics.output.metrics.component_sent_events_total
+    	component_sent_event_bytes_total:    components.sources.internal_metrics.output.metrics.component_sent_event_bytes_total
+    	component_sent_bytes_total:          components.sources.internal_metrics.output.metrics.component_sent_bytes_total
 		events_discarded_total:              components.sources.internal_metrics.output.metrics.events_discarded_total
 		processing_errors_total:             components.sources.internal_metrics.output.metrics.processing_errors_total
 		kafka_queue_messages:                components.sources.internal_metrics.output.metrics.kafka_queue_messages
